Extract numeric constraint tests into a dedicated ExclusiveLimitSerializerTest class
- Moved tests for numeric constraints and exclusive bounds from `SchemaTest` to `ExclusiveLimitSerializerTest`.
- Added additional test cases for invalid exclusive bound values in JSON and YAML.
- Updated test assertions for better coverage of exclusive bound serialization and deserialization edge cases.
- Refined mappings of `SchemaNumberConstraints` to `Constraints.Number` with additional test cases.
